[CSM Portal] Rename sftpgo-authentication-service module path to match integrations/ location
