+++ integrations/sftpgo-authentication-service/internal/handler/handler_test.go
@@ -24,11 +24,11 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/wso2-open-operations/cs-tools/operations/sftpgo-authentication-service/internal/config"
-	"github.com/wso2-open-operations/cs-tools/operations/sftpgo-authentication-service/internal/constants"
-	"github.com/wso2-open-operations/cs-tools/operations/sftpgo-authentication-service/internal/log"
-	"github.com/wso2-open-operations/cs-tools/operations/sftpgo-authentication-service/internal/models"
-	"github.com/wso2-open-operations/cs-tools/operations/sftpgo-authentication-service/internal/service"
+	"github.com/wso2-open-operations/cs-tools/integrations/sftpgo-authentication-service/internal/config"
+	"github.com/wso2-open-operations/cs-tools/integrations/sftpgo-authentication-service/internal/constants"
+	"github.com/wso2-open-operations/cs-tools/integrations/sftpgo-authentication-service/internal/log"
+	"github.com/wso2-open-operations/cs-tools/integrations/sftpgo-authentication-service/internal/models"
+	"github.com/wso2-open-operations/cs-tools/integrations/sftpgo-authentication-service/internal/service"
 )
 
 func TestHandler_Authenticate(t *testing.T) {
