+++ integrations/sftpgo-authentication-service/internal/service/idp_test.go
@@ -23,8 +23,8 @@ import (
 	"net/url"
 	"testing"
 
-	"github.com/wso2-open-operations/cs-tools/operations/sftpgo-authentication-service/internal/config"
-	"github.com/wso2-open-operations/cs-tools/operations/sftpgo-authentication-service/internal/log"
+	"github.com/wso2-open-operations/cs-tools/integrations/sftpgo-authentication-service/internal/config"
+	"github.com/wso2-open-operations/cs-tools/integrations/sftpgo-authentication-service/internal/log"
 )
 
 func TestIsInternalUser(t *testing.T) {
